Make first my homepage

+++ app/controllers/foods_controller.rb
@@ -1,4 +1,6 @@
 class FoodsController < ApplicationController
+	before_action :login_check
+	skip_before_action :login_check, :only => [:posts, :posts_category, :show]
   def posts
 		@posts = Post.all
   end
@@ -19,6 +21,7 @@ class FoodsController < ApplicationController
 
   def show
 		@post = Post.find(params[:id])
+		@comment_writer = User.where(id: session[:user_id])[0]
   end
 
   def write
@@ -26,9 +29,11 @@ class FoodsController < ApplicationController
 
   def write_complete
 		post = Post.new
+		post.user_id = session[:user_id]
 		post.category = params[:post_category]
 		post.title = params[:post_title]
 		post.content = params[:post_content]
+		post.image = params[:image]
 		if post.save
 			flash[:alert] = "저장되었습니다."
 			redirect_to "/foods/show/#{post.id}"
@@ -36,10 +41,14 @@ class FoodsController < ApplicationController
 			flash[:alert] = post.errors.values.flatten.join(' ')
 			redirect_to :back
 		end
-  end
+ end
 
   def edit
 		@post = Post.find(params[:id])
+		if @post.user_id != session[:user_id]
+			flash[:alert] = "수정 권한이 없습니다."
+			redirect_to :back
+		end
   end
 
   def edit_complete
@@ -58,13 +67,19 @@ class FoodsController < ApplicationController
 
   def delete_complete
 		post = Post.find(params[:id])
-		post.destroy
-		flash[:alert] = "삭제되었습니다."
-		redirect_to "/"
+		if post.user_id == session[:user_id]
+			post.destroy
+			flash[:alert] = "삭제되었습니다."
+			redirect_to "/"
+		else
+			flash[:alert] = "삭제 권한이 없습니다."
+			redirect_to :back
+		end
   end
 
 	def write_comment_complete
 		comment = Comment.new
+		comment.user_id = session[:user_id]
 		comment.post_id = params[:post_id]
 		comment.content = params[:comment_content]
 		comment.save
@@ -75,8 +90,13 @@ class FoodsController < ApplicationController
 
 	def delete_comment_complete
 		comment = Comment.find(params[:id])
-		comment.destroy
-		flash[:alert] = "댓글이 삭제되었습니다."
-		redirect_to "/foods/show/#{comment.post_id}"
+		if comment.user_id == session[:user_id]
+			comment.destroy
+			flash[:alert] = "댓글이 삭제되었습니다."
+			redirect_to "/foods/show/#{comment.post_id}"
+		else
+			flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+			redirect_to :back
+		end
 	end
 end
